+++ src/core/model/assets/pending_asset.hpp
@@ -11,6 +11,7 @@ struct PendingAsset
     int id = 0;
     QUrl url;
     QByteArray data;
+    QString name_alias;
     bool loaded = false;
 };