+++ src/core/model/font/custom_font.cpp
@@ -1,6 +1,7 @@
 #include "custom_font.hpp"
 
 #include <unordered_map>
+
 #include <QFontDatabase>
 #include <QCryptographicHash>
 
@@ -18,6 +19,11 @@ public:
         data(data)
     {}
 
+    QString family_name() const
+    {
+        return font.familyName();
+    }
+
 
     QRawFont font;
     int database_index = -1;
@@ -25,6 +31,7 @@ public:
     QByteArray data;
     QString source_url;
     QString css_url;
+    std::set<QString> name_aliases;
 };
 
 class glaxnimate::model::CustomFontDatabase::Private
@@ -34,8 +41,28 @@ public:
     // we keep track of hashes to avoid registering the exact same file twice
     std::unordered_map<QByteArray, int> hashes;
 
+    std::unordered_map<QString, std::vector<int>> name_aliases;
+
+    void tag_alias(const DataPtr& data, const QString& name)
+    {
+        if ( !name.isEmpty() && data->name_aliases.insert(name).second )
+            name_aliases[name].push_back(data->database_index);
+    }
+
     void uninstall(std::unordered_map<int, DataPtr>::iterator iterator)
     {
+        for ( const auto& name : iterator->second->name_aliases )
+        {
+            auto iter = name_aliases.find(name);
+            if ( iter != name_aliases.end() )
+            {
+                if ( iter->second.size() <= 1 )
+                    name_aliases.erase(iter);
+                else
+                    iter->second.erase(std::find(iter->second.begin(), iter->second.end(), iterator->second->database_index));
+            }
+        }
+
         hashes.erase(iterator->second->data_hash);
         QFontDatabase::removeApplicationFont(iterator->first);
         fonts.erase(iterator);
@@ -51,12 +78,16 @@ public:
             uninstall(it);
     }
 
-    DataPtr install(const QByteArray& data)
+    DataPtr install(const QString& name_alias, const QByteArray& data)
     {
         auto hash = QCryptographicHash::hash(data, QCryptographicHash::Sha1);
         auto hashit = hashes.find(hash);
         if ( hashit != hashes.end() )
-            return fonts.at(hashit->second);
+        {
+            auto item = fonts.at(hashit->second);
+            tag_alias(item, name_alias);
+            return item;
+        }
 
 
         QRawFont raw(data, 16);
@@ -71,6 +102,7 @@ public:
 
         auto ptr = std::make_shared<CustomFontData>(raw, index, hash, data);
         fonts.emplace(index, ptr);
+        tag_alias(ptr, name_alias);
         return ptr;
     }
 };
@@ -99,9 +131,9 @@ std::vector<glaxnimate::model::CustomFont> glaxnimate::model::CustomFontDatabase
     return fonts;
 }
 
-glaxnimate::model::CustomFont glaxnimate::model::CustomFontDatabase::add_font(const QByteArray& ttf_data)
+glaxnimate::model::CustomFont glaxnimate::model::CustomFontDatabase::add_font(const QString& name_alias, const QByteArray& ttf_data)
 {
-    return d->install(ttf_data);
+    return d->install(name_alias, ttf_data);
 }
 
 glaxnimate::model::CustomFont glaxnimate::model::CustomFontDatabase::get_font(int database_index)
@@ -111,6 +143,50 @@ glaxnimate::model::CustomFont glaxnimate::model::CustomFontDatabase::get_font(in
         return {};
     return it->second;
 }
+
+QFont glaxnimate::model::CustomFontDatabase::font(const QString& family, const QString& style_name, qreal size) const
+{
+    auto it = d->name_aliases.find(family);
+    if ( it == d->name_aliases.end() )
+    {
+        QFont font(family);
+        font.setPointSizeF(size);
+        font.setStyleName(style_name);
+        return font;
+    }
+
+    CustomFontData* match = d->fonts.at(it->second[0]).get();
+    for ( int id : it->second )
+    {
+        const auto& font = d->fonts.at(id);
+        if ( font->font.styleName() == style_name )
+        {
+            match = font.get();
+            break;
+        }
+    }
+
+    QFont font(match->family_name());
+    font.setPointSizeF(size);
+    font.setStyleName(style_name);
+    return font;
+}
+
+std::unordered_map<QString, std::set<QString>> glaxnimate::model::CustomFontDatabase::aliases() const
+{
+    std::unordered_map<QString, std::set<QString>> map;
+
+    for ( const auto& p : d->name_aliases )
+    {
+        std::set<QString> names;
+        for ( const auto& id : p.second )
+            names.insert(d->fonts.at(id)->family_name());
+        map[p.first] = names;
+    }
+
+    return map;
+}
+
 
 glaxnimate::model::CustomFont::CustomFont(CustomFontDatabase::DataPtr dd)
     : d(std::move(dd))
@@ -126,11 +202,6 @@ glaxnimate::model::CustomFont::CustomFont()
 
 glaxnimate::model::CustomFont::CustomFont(int database_index)
     : CustomFont(CustomFontDatabase::instance().get_font(database_index))
-{
-}
-
-glaxnimate::model::CustomFont::CustomFont(const QByteArray& data)
-    : CustomFont(CustomFontDatabase::instance().d->install(data))
 {
 }
 
@@ -159,7 +230,7 @@ int glaxnimate::model::CustomFont::database_index() const
 
 QString glaxnimate::model::CustomFont::family() const
 {
-    return d->font.familyName();
+    return d->family_name();
 }
 
 QString glaxnimate::model::CustomFont::style_name() const